Update app SQL relationships

+++ app/routers/post.py
@@ -12,9 +12,13 @@ router = APIRouter(
     tags=["Posts"]
 )
 
-@router.get("/")
-def sqlalchemy_test(db: Session = Depends(get_db), response_model=list[schemas.Post]):
+@router.get("/", response_model=list[schemas.Post])
+def sqlalchemy_test(db: Session = Depends(get_db), current_user: int = Depends (oauth2.get_current_user)):
     posts = db.query(models.Post).all()
+    print(current_user.id)
+    ## Query with filter
+    #posts = db.query(models.Post).filter(models.Post.user_id == current_user.id).all()
+
     ## ==> SELECT * FROM posts
     print(posts)
     return  posts
@@ -23,8 +27,8 @@ def sqlalchemy_test(db: Session = Depends(get_db), response_model=list[schemas.P
 
 
 ## Get a specific post
-@router.get("/{id}", response_model=schemas.Post, )
-async def get_post(id: int, db: Session = Depends(get_db)):
+@router.get("/{id}", response_model=schemas.Post)
+async def get_post(id: int, db: Session = Depends(get_db), current_user: int = Depends (oauth2.get_current_user)):
 
     ## Using SQLAlchemy ORM
     post = db.query(models.Post).filter(models.Post.id == id).first()
@@ -52,11 +56,11 @@ async def get_post(id: int, db: Session = Depends(get_db)):
 
 
 @router.post("/", status_code=status.HTTP_201_CREATED, response_model=schemas.Post)
-def create_post(post: schemas.PostCreate, db: Session = Depends(get_db), user_id: int = Depends (oauth2.get_current_user)):
-
+def create_post(post: schemas.PostCreate, db: Session = Depends(get_db), current_user: int = Depends (oauth2.get_current_user)):
+    print(current_user.id)
 ## Using SQLAlchemy ORM
     #new_post = models.Post(title=post.title, content=post.content, published=post.published)
-    new_post = models.Post(**post.model_dump())  # unpacking the post object
+    new_post = models.Post(user_id=current_user.id, **post.model_dump())  # unpacking the post object
     ## Add the new post to the session
     db.add(new_post)
     db.commit()
@@ -73,7 +77,7 @@ def create_post(post: schemas.PostCreate, db: Session = Depends(get_db), user_id
 
 ### Delete a post ###
 @router.delete("/{id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post(id: int, db: Session = Depends(get_db), user_id: int = Depends (oauth2.get_current_user)):
+async def delete_post(id: int, db: Session = Depends(get_db), current_user: int = Depends (oauth2.get_current_user)):
 
     deleted_rows = (
         db.query(models.Post)
@@ -85,13 +89,17 @@ async def delete_post(id: int, db: Session = Depends(get_db), user_id: int = Dep
                 status_code=status.HTTP_404_NOT_FOUND,
                 detail=f"Post with id: {id} does not exist"
                 )
+    ## if user is not owner of the post == not authorized
+    if deleted_rows.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to perform requested action")
+
     db.delete(deleted_rows)
     db.commit()
     return Response(status_code=status.HTTP_204_NO_CONTENT)
 
 ### Update a post ###
 @router.put("/{id}")
-def update_post(id: int, post: schemas.PostCreate, db: Session = Depends(get_db), response_model=schemas.Post, user_id: int = Depends (oauth2.get_current_user)):
+def update_post(id: int, post: schemas.PostCreate, db: Session = Depends(get_db), response_model=schemas.Post, current_user: int = Depends (oauth2.get_current_user)):
     #
     updated_post = (
         db.query(models.Post)
@@ -104,6 +112,10 @@ def update_post(id: int, post: schemas.PostCreate, db: Session = Depends(get_db)
                 status_code=status.HTTP_404_NOT_FOUND,
                 detail=f"Post with id: {id} does not exist"
                 )
+    # if user is not owner of the post == not authorized
+    if updated_post.user_id != current_user.id:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to perform requested action")
+
     # mettre à jour les attributs
     #updated_post.title = post.title
     updated_post.content = post.content
